Code BE : update trainee eval , milestone eval

+++ BE/online-learning-system/src/main/java/swp490/g23/onlinelearningsystem/entities/milestone_eval/domain/MilestoneEval.java
@@ -33,7 +33,7 @@ public class MilestoneEval extends BaseEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long milestoneEvalId;
 
-    private Long bonus;
+    private Double bonus;
 
     private Double grade;
